waitNextFrame: cast the OSCheckStopwatch (s64 OSTime) result to u64 so the elapsed-tick-to-float conversion goes through __cvt_ull_flt as retail does, not __cvt_sll_flt. A real signed-vs-unsigned semantic difference that every gate scored as clean: the function was already fuzzy 100.0 because objdiff does not compare the callee symbol of a REL24 relocation. With the cast the function is byte-identical to retail apart from anon-vs-named .sdata2 pool naming. Unit main/main/pi_dolphin 99.1836 and matched_data 8472 both unchanged; unit is linked False so main.dol is bit-identical e750e8 size 3398976

+++ src/main/pi_dolphin.c
@@ -4639,7 +4639,8 @@ void waitNextFrame(void)
     u32 frames;
 
     OSStopStopwatch(&lbl_8035F680);
-    lbl_803DCCC0 = OSCheckStopwatch(&lbl_8035F680) / (f32)(u32)((*(u32*)0x800000f8 >> 2) / 1000);
+    lbl_803DCCC0 =
+        (u64)OSCheckStopwatch(&lbl_8035F680) / (f32)(u32)((*(u32*)0x800000f8 >> 2) / 1000);
     OSResetStopwatch(&lbl_8035F680);
     OSStartStopwatch(&lbl_8035F680);
     timeDelta = physicsTimeScale * (lbl_803DEAA0 * lbl_803DCCC0);
